refactor(security): unify mask_secrets into shared sanitize module
Extract mask_secrets from observe.rs into shared/sanitize.rs with
broader coverage (Bearer, sk-*, password/token/apikey/secret/private_key).
Both observe and orchestrate modules now use the same regex set,
eliminating the coverage gap where run.json could store secrets that
observe would have masked.
